♻️ refactor: Types/interfaces exportadas separadas
