+++ src/@types/bookSimpleData.ts
@@ -1,0 +1,10 @@
+import { SellerData } from './sellerData'
+
+export interface BookSimpleData {
+  id: string
+  name: string
+  author: string
+  shortDescription: string
+  category: string
+  seller: SellerData
+}
