+++ src/@types/sellerData.ts
@@ -1,0 +1,6 @@
+export interface SellerData {
+  name: string
+  location: string
+  averageRating: number
+  avaliationsNumber: number
+}
