+++ src/hooks/useSingleBook.ts
@@ -1,26 +1,6 @@
+import { BookCompleteData } from '@/@types/bookCompleteData'
 import { api } from '@/lib/axios'
 import { useQuery } from '@tanstack/react-query'
-
-interface Seller {
-  name: string
-  location: string
-  averageRating: number
-  avaliationsNumber: number
-  phoneNumber: string
-}
-
-export interface BookCompleteData {
-  id: string
-  name: string
-  author: string
-  category: string
-  language: string
-  year: string
-  publishingCompany: string
-  state: string
-  description: string
-  seller: Seller
-}
 
 async function fetchBookById(id: string) {
   const response = await api.get(`/books-details/${id}`)
